pyproject.toml
[tool.poetry]
name = "e-shop-drf"
version = "0.1.0"
description = ""
authors = ["Sumerian <shemerov.s@yandex.ru>"]
readme = "README.md"
packages = [{include = "e_shop_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
